+++ Changelog.md
@@ -6,6 +6,8 @@ Released on ?.
 
 ✨ New Features / Improvements
 
+- Height profile now shows track name and an info button to go to the track
+  details page
 - Using CesiumJS version 1.122.0
 
 🐛 Bugfixes